% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_dataset}
\alias{example_dataset}
\title{Example Dataset for MOCHIS DEG Analysis}
\format{
This dataset contains the following objects:
\describe{
\item{example_data1}{A numeric matrix of normalized counts for all age
groups.}
\item{labels1}{A vector of sample labels for all age groups.}
\item{example_data2}{A numeric matrix of normalized counts for age groups
20-39 and 60-79.}
\item{labels2}{A vector of sample labels for age groups 20-39 and 60-79.}
}
}
\source{
Normalized counts from GTEx for package demonstration purposes.
}
\usage{
example_dataset
}
\description{
This dataset contains a subset of rounded normalized gene expression counts
from GTEx whole blood tissue and associated labels for demonstration
purposes.
}
\examples{
data(example_dataset)
}
\keyword{datasets}
