% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_dataset_raw_3000_genes}
\alias{example_dataset_raw_3000_genes}
\title{Example Dataset for MOCHIS DEG Analysis with Raw Counts}
\format{
This dataset contains a list with the following components
\describe{
\item{COUNTS}{A numeric matrix of raw counts for 3000 genes.}
\item{METADATA}{The metadata for the corresponding raw counts.}
}
}
\source{
Generated for package documentation
}
\usage{
example_dataset_raw_3000_genes
}
\description{
This dataset contains raw gene expression counts from 3000 genes in GTEx whole blood tissue for demonstration purposes.
}
\examples{
data(example_dataset_raw_3000_genes)
}
\keyword{datasets}
