% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_analysis.R
\name{QRscore.genetest}
\alias{QRscore.genetest}
\title{Perform Differential Expression Analysis using QRscore}
\usage{
QRscore.genetest(
  normalized_matrix,
  labels,
  pairwise_test = FALSE,
  pairwise_logFC = FALSE,
  test_mean = TRUE,
  test_dispersion = FALSE,
  num_cores = 1,
  silent = TRUE,
  rank = TRUE,
  ...
)
}
\arguments{
\item{normalized_matrix}{A prefiltered normalized matrix (gene x sample).}

\item{labels}{A vector of labels corresponding to the samples.}

\item{pairwise_test}{Logical, whether to perform pairwise test statistics for more than two groups.}

\item{pairwise_logFC}{Logical, whether to calculate pairwise log fold changes for mean and variance.}

\item{test_mean}{Logical, whether to test the mean.}

\item{test_dispersion}{Logical, whether to test the dispersion.}

\item{num_cores}{Integer, number of cores to use for parallel processing.}

\item{silent}{Logical, whether to suppress all messages except warnings and errors.}

\item{...}{Additional arguments passed to the \code{QRscore.test} function.}
}
\value{
A list with two data frames: \code{mean_test} and \code{var_test}.
}
\description{
This function performs differential expression analysis using the QRscore method.
}
\examples{
data(example_dataset)
results <- QRscore.genetest(example_dataset$example_data1, example_dataset$labels1, 
                               pairwise_test = TRUE, pairwise_logFC = TRUE, 
                               test_mean = TRUE, test_dispersion = TRUE, num_cores = 4,
                               approx = "asymptotic")
                             
head(results$mean_test)

results2 <- QRscore.genetest(example_dataset$example_data2, example_dataset$labels2, 
                               pairwise_test = TRUE, pairwise_logFC = TRUE, 
                               test_mean = TRUE, test_dispersion = TRUE, num_cores = 4,
                               approx = "asymptotic")
head(results2$var_test)
}
